Added Main Program
Currently, the program uses hijacks the GfxTablet app on F-Droid, and a custom keyboard shortcut on Libre-Office Draw (CTRL + W) for the freeline tool, to write whenever the user does. This, of course, can and will be modified, and the program will likely be based around Krita. 

The program also uses LTO by default, though the performance gains from this are negligible, it's still useful if you want the program to be as responsive as possible.